+++ src/lib/decorate.js
@@ -1,12 +1,12 @@
 const addDots = require('./dotPosition');
 const addAbbreviations = require('./abbreviate');
 const addTimestamps = require('./timestamps');
-const addKeyDevelopments = require('./editorial-lists');
+const addKeyDevelopments = require('./key-developments');
 
 module.exports = function decorate (data, keyDevelopments) {
 	if (keyDevelopments.length) {
-		const dataWithKeyDevelopements = addKeyDevelopments(data, keyDevelopments);
-		return addTimestamps(addAbbreviations(addDots(dataWithKeyDevelopements)));
+		const dataWithKeyDevelopments = addKeyDevelopments(data, keyDevelopments);
+		return addTimestamps(addAbbreviations(addDots(dataWithKeyDevelopments)));
 	} else {
 		return addTimestamps(addAbbreviations(addDots(data)));
 	}
